Admin Charts-Working

+++ CustomGuitars/order/models.py
@@ -9,7 +9,7 @@ from accounts.models import CustomUser
 class Order(models.Model):
  user = models.ForeignKey(CustomUser, on_delete=models.CASCADE, null=True,)
  token = models.CharField(max_length=250, blank=True) 
- total = models.DecimalField(max_digits=10, decimal_places=2, verbose_name='Euro Order Total') 
+ total = models.IntegerField(verbose_name='Euro Order Total') 
  emailAddress = models.EmailField(max_length=250, blank=True, verbose_name='Email Address') 
  created = models.DateTimeField(auto_now_add=True) 
  billingName = models.CharField(max_length=250, blank=True)
